+++ Tests/KeyboardKitTests/Layout/Providers/StandardKeyboardLayoutProviderTests.swift
@@ -12,47 +12,48 @@ import XCTest
 class StandardKeyboardLayoutProviderTests: XCTestCase {
     
     var inputSetProvider: MockInputSetProvider!
-    var context: KeyboardContext!
+    var keyboardContext: KeyboardContext!
     var provider: StandardKeyboardLayoutProvider!
 
     override func setUp() {
-        context = KeyboardContext()
+        keyboardContext = KeyboardContext()
         inputSetProvider = MockInputSetProvider()
         inputSetProvider.alphabeticInputSetValue = AlphabeticInputSet(rows: [["a", "b", "c"], ["a", "b", "c"], ["a", "b", "c"]].map(InputSetRow.init(chars:)))
         inputSetProvider.numericInputSetValue = NumericInputSet(rows: [["1", "2", "3"], ["1", "2", "3"], ["1", "2", "3"]].map(InputSetRow.init(chars:)))
         inputSetProvider.symbolicInputSetValue = SymbolicInputSet(rows: [[",", ".", "-"], [",", ".", "-"], [",", ".", "-"]].map(InputSetRow.init(chars:)))
         provider = StandardKeyboardLayoutProvider(
+            keyboardContext: keyboardContext,
             inputSetProvider: inputSetProvider,
             dictationReplacement: .primary(.go))
     }
 
     func testKeyboardLayoutProviderForContextIsPhoneProviderIfContextDeviceIsPhone() {
-        context.deviceType = .phone
-        let result = provider.keyboardLayoutProvider(for: context)
+        keyboardContext.deviceType = .phone
+        let result = provider.keyboardLayoutProvider(for: keyboardContext)
         XCTAssertTrue(result === provider.iPhoneProvider)
     }
 
     func testKeyboardLayoutProviderForContextIsPhoneProviderIfContextDeviceIsPad() {
-        context.deviceType = .pad
-        let result = provider.keyboardLayoutProvider(for: context)
+        keyboardContext.deviceType = .pad
+        let result = provider.keyboardLayoutProvider(for: keyboardContext)
         XCTAssertTrue(result === provider.iPadProvider)
     }
 
 
     func testKeyboardLayoutForContextIsPhoneLayoutIfContentDeviceIsPhone() {
-        context.deviceType = .phone
-        let layout = provider.keyboardLayout(for: context)
-        let phoneLayout = provider.iPhoneProvider.keyboardLayout(for: context)
-        let padLayout = provider.iPadProvider.keyboardLayout(for: context)
+        keyboardContext.deviceType = .phone
+        let layout = provider.keyboardLayout(for: keyboardContext)
+        let phoneLayout = provider.iPhoneProvider.keyboardLayout(for: keyboardContext)
+        let padLayout = provider.iPadProvider.keyboardLayout(for: keyboardContext)
         XCTAssertEqual(layout.itemRows, phoneLayout.itemRows)
         XCTAssertNotEqual(layout.itemRows, padLayout.itemRows)
     }
 
     func testKeyboardLayoutForContextIsPadLayoutIfContentDeviceIsPad() {
-        context.deviceType = .pad
-        let layout = provider.keyboardLayout(for: context)
-        let phoneLayout = provider.iPhoneProvider.keyboardLayout(for: context)
-        let padLayout = provider.iPadProvider.keyboardLayout(for: context)
+        keyboardContext.deviceType = .pad
+        let layout = provider.keyboardLayout(for: keyboardContext)
+        let phoneLayout = provider.iPhoneProvider.keyboardLayout(for: keyboardContext)
+        let padLayout = provider.iPadProvider.keyboardLayout(for: keyboardContext)
         XCTAssertNotEqual(layout.itemRows, phoneLayout.itemRows)
         XCTAssertEqual(layout.itemRows, padLayout.itemRows)
     }
@@ -60,7 +61,7 @@ class StandardKeyboardLayoutProviderTests: XCTestCase {
 
     func testRegisteringInputSetProviderChangesInstanceForAllProviders() {
         let newProvider = MockInputSetProvider()
-        context.deviceType = .phone
+        keyboardContext.deviceType = .phone
         provider.register(inputSetProvider: newProvider)
         XCTAssertFalse(provider.inputSetProvider === inputSetProvider)
         XCTAssertTrue(provider.inputSetProvider === newProvider)